Better version of yesterday's - Tested by sampling 4 accelerometers on same spi line. Able to do this at 12.8kHz without the buffer filling up in between. Previously I had accidentally been starting the sample at the wrong address (0x3C instead of 0x3F), so it was reading the data but I guess not registering that data was being cleared from the buffer? Not sure.

+++ Payload_2026_Flight_Accel/Core/Inc/main.h
@@ -60,7 +60,7 @@ extern uint8_t rx_buff_stat[2];
 
 extern uint16_t curr_frame;
 extern uint8_t curr_accelerometer;
-extern uint8_t bytes_to_read;
+extern uint16_t bytes_to_read;
 extern uint8_t *write_address;
 extern uint8_t *read_address;
